+++ packages/recommend-core/src/__tests__/getRecommendations.test.ts
@@ -1,69 +1,30 @@
-import {
-  createMultiSearchResponse,
-  createRecommendClient,
-} from '../../../../test/utils';
+import { createRecommendationsClient } from '../../../../test/utils';
 import { getRecommendations } from '../getRecommendations';
 import { RecommendModel } from '../types';
 
-const hit = {
-  name: 'Landoh 4-Pocket Jumpsuit',
-  category: 'Women - Jumpsuits-Overalls',
-  price: 250,
-  url: 'women/jumpsuits-overalls/d06270-9132-995',
-  hierarchical_categories: {
-    lvl0: 'women',
-    lvl1: 'women > jeans & bottoms',
-    lvl2: 'women > jeans & bottoms > jumpsuits & overalls',
-  },
-  keywords: [
-    'women',
-    'jeans & bottoms',
-    'jumpsuits & overalls',
-    'Jumpsuits',
-    'Loose',
-    'Woven',
-    'Long sleeve',
-    'Grey',
-  ],
-  objectID: 'D06270-9132-995',
-  recommendations: [
-    { objectID: '1', score: 1.99 },
-    { objectID: '2', score: 2.99 },
-    { objectID: '3', score: 3.99 },
-  ],
-};
+describe('getRecommendations', () => {
+  test('does not forward unwanted props', async () => {
+    const { recommendClient } = createRecommendationsClient();
+    const props = {
+      model: 'related-products' as RecommendModel,
+      recommendClient,
+      indexName: 'indexName',
+      objectIDs: ['objectID'],
+      a: 'b',
+    };
 
-function createRecommendationsClient() {
-  const recommendClient = createRecommendClient({
-    getFrequentlyBoughtTogether: jest.fn(() =>
-      Promise.resolve(
-        createMultiSearchResponse({
-          hits: [hit],
-        })
-      )
-    ),
-    getRelatedProducts: jest.fn(() =>
-      Promise.resolve(
-        createMultiSearchResponse({
-          hits: [hit],
-        })
-      )
-    ),
-    getRecommendations: jest.fn(() =>
-      Promise.resolve(
-        createMultiSearchResponse({
-          hits: [hit],
-        })
-      )
-    ),
+    await getRecommendations(props);
+
+    expect(recommendClient.getRecommendations).toHaveBeenCalledTimes(1);
+    expect(recommendClient.getRecommendations).toHaveBeenCalledWith([
+      {
+        model: 'related-products',
+        indexName: 'indexName',
+        objectID: 'objectID',
+      },
+    ]);
   });
 
-  return {
-    recommendClient,
-  };
-}
-
-describe('getRecommendations', () => {
   test('calls the correct index for "related-products"', async () => {
     const { recommendClient } = createRecommendationsClient();
     const props = {
