+++ test/utils/index.ts
@@ -1,2 +1,3 @@
 export * from './createApiResponse';
+export * from './createRecommendationsClient';
 export * from './createRecommendClient';
